+++ app.R
@@ -235,4 +235,3 @@ server <- function(input, output, session) {
 }
 
 shinyApp(ui, server)
-